+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## 2.1.1 under development
 
+- New #92: A third optional argument `$attributes` containing tag attributes in terms of name-value pairs has been
+  added to methods `Html::textInput()`, `Html::hiddenInput()`, `Html::passwordInput()`, `Html::fileInput()`,
+  `Html::radio()`, `Html::checkbox()`, `Html::textarea()` (vjik)
 - New #89: Add method `nofollow()` to the `A` tag (soodssr)
 
 ## 2.1.0 September 23, 2021